fixing errors
Venues.js. - add change venue name if it is meaningless. Array.some() method is used to iterate through the array of words and check if any of the words are included in the lowercase version of the name string. If at least one of the words is found, it changes the name to 'Accommodation'.

UserVenues.js - fetchBookingDetails was put in callBack() - to ensure that the function is memoized and recreated only when those dependencies change.